+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-All notable changes to the Wazuh app for Splunk project will be documented in this file.
+All notable changes to the Wazuh App for Splunk project will be documented in this file.
 
 ## Wazuh v4.3.4 - Splunk Enterprise v8.1.[1-10], v8.2.x - Revision 4307